requirements.txt
fastapi[standard]
uvicorn[standard]
pydantic
pydantic-settings
supabase
PyJWT
python-dotenv
apscheduler
httpx
brevo-python